Refactor: Harden Project with Stricter Linting and Component Refactoring
This commit introduces a series of improvements to harden the project, improve code quality, and ensure long-term maintainability.

Key changes:
- **Stricter Linting:** The ESLint configuration has been updated with stricter rules, including warnings for explicit `any`, strict boolean expressions, and sorted imports. This will help catch potential bugs and enforce a consistent code style.
- **Admin Dashboard Refactoring:** The `AgentList` component has been refactored to be more modular and maintainable. The `AgentForm` has been extracted into its own file, and the core logic has been moved into three new custom hooks: `useAgentManagement`, `useAgentFilters`, and `useModelFetcher`.
- **Dependency Audit:** A `DEPENDENCY_AUDIT.md` file has been added to the repository, providing a detailed report on outdated dependencies and a recommended update strategy.
- **State Management Analysis:** Identified that the project is not using `@tanstack/react-query` and is instead relying on a custom context-based solution. A future refactoring to adopt React Query is recommended.
