+++ DEPENDENCY_AUDIT.md
@@ -1,0 +1,61 @@
+# Dependency Audit Report
+
+This report outlines the outdated dependencies in the project and provides recommendations for updating them.
+
+## Summary
+
+The project has a significant number of outdated dependencies, including several with new major versions. Updating these packages will bring new features, performance improvements, and security patches, but it also carries the risk of introducing breaking changes.
+
+A full dependency update, especially for major versions, should be treated as a separate, dedicated project with a thorough testing plan.
+
+## High-Risk Updates (Major Versions)
+
+These packages have new major versions available. Updating them will likely require significant code changes and should be done one at a time, with careful testing after each update.
+
+- **React (`react`, `react-dom`)**: `18.3.1` -> `19.1.1`
+  - **Recommendation:** React 19 introduces many new features and changes. A migration guide should be consulted. This is the highest-risk update and should be planned carefully.
+- **Three.js (`three`)**: `0.149.0` -> `0.179.1`
+  - **Recommendation:** This is a very large jump in versions. The Three.js migration guide should be followed closely. Given the project's reliance on 3D rendering, this update will require extensive testing of all CAD-related features.
+- **Next.js (`next`, `eslint-config-next`)**: `15.3.2` -> `15.4.6`
+  - **Recommendation:** While this is a minor version bump, Next.js updates can sometimes introduce subtle breaking changes. The Next.js blog and release notes should be reviewed before updating.
+- **Tailwind CSS (`tailwindcss`)**: `3.4.17` -> `4.1.11`
+  - **Recommendation:** Tailwind CSS v4 has a new engine and configuration. The migration guide is essential.
+- **Zod (`zod`)**: `3.25.76` -> `4.0.15`
+  - **Recommendation:** Zod is used for schema validation. The release notes should be checked for any breaking changes in the API.
+- **@hookform/resolvers**: `3.10.0` -> `5.2.1`
+  - **Recommendation:** This is a large jump and will likely have breaking changes.
+- **date-fns**: `2.30.0` -> `4.1.0`
+  - **Recommendation:** Check the `date-fns` changelog for any breaking changes to the date formatting and manipulation functions used in the project.
+
+## Medium-Risk Updates (Minor Versions)
+
+These packages have new minor versions available. They should be updated with caution, and the application should be tested after each group of updates.
+
+- **@radix-ui/***: Most `@radix-ui` packages have minor updates available.
+- **@types/***: Many type definition packages can be updated.
+- **lucide-react**: `0.454.0` -> `0.537.0`
+- **sonner**: `1.7.4` -> `2.0.7`
+- **recharts**: `2.15.0` -> `3.1.2`
+- **and many others...**
+
+**Recommendation:** These can likely be updated in batches. For example, all `@radix-ui` packages could be updated together. After each batch update, the application should be tested to ensure that the UI and functionality are still working correctly.
+
+## Low-Risk Updates (Patch Versions)
+
+Many packages have patch updates available. These should be safe to update.
+
+**Recommendation:** Update all patch versions at once. This is unlikely to cause any issues.
+
+## Ignored Packages
+
+- **`three-orbitcontrols`**: `2.110.3` -> `2.110.2`. This appears to be an error in the package registry, as it suggests a downgrade. This should be ignored.
+
+## Next Steps
+
+1.  **Create a dedicated branch** for the dependency updates.
+2.  **Start with the low-risk patch updates.**
+3.  **Move on to the medium-risk minor updates**, updating them in small, logical batches.
+4.  **Tackle the high-risk major updates one by one**, starting with the less critical ones and leaving React, Three.js, and Next.js for last.
+5.  **Thoroughly test the application** at each stage of the update process.
+
+By following this plan, the project's dependencies can be brought up to date in a safe and controlled manner.
