--- hooks/use-mobile.tsx
@@ -1,24 +1,0 @@
-"use client"
-
-import { useState, useEffect } from "react"
-
-export function useMobile() {
-  const [isMobile, setIsMobile] = useState(false)
-
-  useEffect(() => {
-    const checkIfMobile = () => {
-      setIsMobile(window.innerWidth < 768)
-    }
-
-    // Initial check
-    checkIfMobile()
-
-    // Add event listener
-    window.addEventListener("resize", checkIfMobile)
-
-    // Clean up
-    return () => window.removeEventListener("resize", checkIfMobile)
-  }, [])
-
-  return isMobile
-}
