+++ package.json
@@ -117,6 +117,7 @@
     "@types/uuid": "9.0.7",
     "eslint": "^9",
     "eslint-config-next": "15.3.2",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
     "postcss": "^8",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.8.3"
