+++ tsconfig.json
@@ -39,7 +39,8 @@
     "forceConsistentCasingInFileNames": true,
     "declaration": false,
     "declarationMap": false,
-    "sourceMap": false
+    "sourceMap": false,
+    "target": "ES2017"
   },
   "include": [
     "next-env.d.ts",
